Add dev setup for basic cookiecutter
Now, if the user runs their trame app with the `--dev` argument, the reload
button will appear at the bottom of the web browser. Pressing it will cause
the UI file to be reloaded and updated, which allows for the UI to be
developed and updated without closing the program.

This additionally reloads the controller file, which, after reload, will also
call the `on_reload()` function, which is given a function that can be used to
reload other modules.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>