fix(import): don't expose main() at root level
Some trame features such as the `www` generation need to import a specific part of the module,
such as `{{import_name}}.module`. If we expose `main()` at the root level, then this requires
the `www` generation to import the app as well (which it may not be able to do).

Remove the `main()` exposure at the root level, and access the main function
via `{{import_name}}.app` instead.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>